extract all months helper from isracard
